requirements-dev.txt
# Code formatting
black==25.1.0
isort==6.0.1

# Linting
flake8==7.2.0
flake8-bugbear==24.12.12
flake8-comprehensions==3.16.0
flake8-simplify==0.22.0

# Type checking
mypy==1.16.0

# Testing (for future use)
pytest==8.3.5
pytest-cov==6.0.0
pytest-homeassistant-custom-component==0.13.249

